+++ src/templates/MeetingResult/styled.tsx
@@ -1,0 +1,1 @@
+import styled from 'styled-components';
